requirements.txt
asgiref==3.8.1
certifi==2025.11.12
charset-normalizer==3.4.4
dj-database-url==2.3.0
Django==5.1.4
django-anymail[mailgun]>=10.0,<11.0
django-cors-headers==4.6.0
django-on-heroku==1.1.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.11
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.5
schedule==1.2.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.6.2
wheel==0.45.1
whitenoise==6.8.2
